backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
Pillow==10.1.0
transformers==4.36.2
torch==2.1.2
web3==6.11.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2

